ModalCard and Products
Create ModalCard and use it into Products. First this class was to be a modal and show when user click add in product. Now I change my mind and this components has inputs with product name, price and select the taxes (vat).

"taxRateArray" is created to select a taxes, when user create a new product.

TODO: I would try to change this ModalCard (yes, and would be nice to change a name) and use it also into Client and User profile.